+++ Configuration/ConfigurationInterface.php
@@ -24,4 +24,11 @@ interface ConfigurationInterface
      * @return string
      */
     function getAliasName();
+
+    /**
+     * Returns whether multiple annotations of this type are allowed
+     *
+     * @return Boolean
+     */
+    function allowArray();
 }