+++ Configuration/Method.php
@@ -66,4 +66,15 @@ class Method extends ConfigurationAnnotation
     {
         return 'method';
     }
+
+    /**
+     * Only one cache directive is allowed
+     *
+     * @return Boolean
+     * @see ConfigurationInterface
+     */
+    public function allowArray()
+    {
+        return false;
+    }
 }