+++ Tests/EventListener/Fixture/FooControllerCacheAtClassAndMethod.php
@@ -10,11 +10,10 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Cache;
 class FooControllerCacheAtClassAndMethod
 {
     const CLASS_SMAXAGE = 20;
-    const METHOD_SMAXAGE = 15;
-    const METHOD_SECOND_SMAXAGE = 25;
+    const METHOD_SMAXAGE = 25;
 
     /**
-     * @Cache(smaxage="15")
+     * @Cache(smaxage="25")
      */
     public function barAction()
     {